Extract AbstractStringOrBytesUuidFunction base class
IsUuidScalarFunction and ToUuidScalarFunction share identical getFunctionInfo
dispatch logic (STRING_FUNCTION_INFO / BYTES_FUNCTION_INFO switch). Extract
it into AbstractStringOrBytesUuidFunction so subclasses only provide the two
FunctionInfo constants via protected abstract methods.

Addresses review comment: the getFunctionInfo pattern appears multiple times.
