+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/uuid/IsUuidScalarFunction.java
@@ -25,9 +25,7 @@ import org.apache.calcite.sql.type.OperandTypes;
 import org.apache.calcite.sql.type.ReturnTypes;
 import org.apache.calcite.sql.type.SqlTypeFamily;
 import org.apache.pinot.common.function.FunctionInfo;
-import org.apache.pinot.common.function.PinotScalarFunction;
 import org.apache.pinot.common.function.sql.PinotSqlFunction;
-import org.apache.pinot.common.utils.DataSchema.ColumnDataType;
 import org.apache.pinot.spi.annotations.ScalarFunction;
 import org.apache.pinot.spi.utils.UuidUtils;
 
@@ -38,7 +36,7 @@ import org.apache.pinot.spi.utils.UuidUtils;
  * <p>This implementation is stateless and thread-safe.
  */
 @ScalarFunction(names = {"IS_UUID"})
-public class IsUuidScalarFunction implements PinotScalarFunction {
+public class IsUuidScalarFunction extends AbstractStringOrBytesUuidFunction {
   private static final FunctionInfo STRING_FUNCTION_INFO;
   private static final FunctionInfo BYTES_FUNCTION_INFO;
 
@@ -53,6 +51,16 @@ public class IsUuidScalarFunction implements PinotScalarFunction {
     } catch (NoSuchMethodException e) {
       throw new RuntimeException(e);
     }
+  }
+
+  @Override
+  protected FunctionInfo getStringFunctionInfo() {
+    return STRING_FUNCTION_INFO;
+  }
+
+  @Override
+  protected FunctionInfo getBytesFunctionInfo() {
+    return BYTES_FUNCTION_INFO;
   }
 
   @Override
@@ -71,28 +79,6 @@ public class IsUuidScalarFunction implements PinotScalarFunction {
     return new PinotSqlFunction("IS_UUID", ReturnTypes.BOOLEAN,
         OperandTypes.or(OperandTypes.family(List.of(SqlTypeFamily.CHARACTER)),
             OperandTypes.family(List.of(SqlTypeFamily.BINARY))));
-  }
-
-  @Nullable
-  @Override
-  public FunctionInfo getFunctionInfo(ColumnDataType[] argumentTypes) {
-    if (argumentTypes.length != 1) {
-      return null;
-    }
-    switch (argumentTypes[0]) {
-      case STRING:
-        return STRING_FUNCTION_INFO;
-      case BYTES:
-        return BYTES_FUNCTION_INFO;
-      default:
-        return null;
-    }
-  }
-
-  @Nullable
-  @Override
-  public FunctionInfo getFunctionInfo(int numArguments) {
-    return numArguments == 1 ? STRING_FUNCTION_INFO : null;
   }
 
   public static boolean isUuid(String value) {
